requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.0

# ML Models - Use versions compatible with Python 3.13
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib==1.4.2
lightgbm==4.1.0
xgboost==2.0.3
lime==0.2.0.1

# Visual Detection (Optional - for image analysis)
torch>=2.6.0
torchvision>=0.20.0
transformers>=4.35.0
Pillow>=10.1.0
opencv-python>=4.8.0
google-cloud-vision==3.4.5
grpcio==1.76.0
grpcio-status==1.62.3

# Text Processing
nltk==3.8.1

# Utilities
requests==2.31.0
